Complete rewrite
It works. It's simple. It removes pretty much every feature it
originally had.

But I couldn't leave this completely broken.

So I made it simpler, and now it can just sit on npm forever and I don't
have to worry about it.

Removes encryption, removes tests, etc etc.

Now it just does the very basics.

And it works.

Which wasn't the case in v1.3

So ... yeah.